requirements.txt
flask==3.0.0
torch==2.4.0
transformers==4.44.2
peft==0.12.0
openai-whisper
accelerate
Pillow
numpy<2.0.0
gunicorn
streamlit
gradio